Fix bug if multiple sticky present and Restore env.

+++ bot/cogs/sticky.py
@@ -18,13 +18,14 @@ class Sticky(commands.GroupCog, group_name="sticky"):
             res = await conn.fetch(
                 "SELECT channel_id,message_id,message FROM sticky WHERE channel_id = $1", message.channel.id
             )
-        if not message.author.bot and res:
-            data = dict(res[0]) 
-            sticky = await message.channel.fetch_message(data['message_id'])
-            await sticky.delete()
-            msg = await message.channel.send(data['message'])
-            async with self.db_pool.acquire() as conn:
-                await conn.executemany("INSERT INTO sticky(channel_id, message_id) VALUES ($1,$2) ON CONFLICT (channel_id) DO UPDATE SET message_id = excluded.message_id", [(message.channel.id, msg.id)])
+            if not message.author.bot and res:
+                data = dict(res[0])
+                if message.channel.id == data['channel_id']:
+                    sticky = await message.channel.fetch_message(data['message_id'])
+                    await sticky.delete()
+                    msg = await message.channel.send(data['message'])
+                    async with self.db_pool.acquire() as conn:
+                        await conn.executemany("INSERT INTO sticky(channel_id, message_id) VALUES ($1,$2) ON CONFLICT (channel_id) DO UPDATE SET message_id = excluded.message_id", [(message.channel.id, msg.id)])
 
     @commands.guild_only()
     @app_commands.command(name='add-sticky', description='Add sticky message to channel')
@@ -36,33 +37,46 @@ class Sticky(commands.GroupCog, group_name="sticky"):
         channel: Union[discord.TextChannel, discord.ForumChannel, discord.Thread]
     ) -> None:
         await interaction.response.defer()
-        async with self.db_pool.acquire() as conn:
-            res = await conn.fetch(
-                "SELECT channel_id FROM sticky WHERE channel_id = $1", channel.id
-            )
-        if not res:
-            target = self.bot.get_channel(channel.id)
-            id = await target.send(message)
-            
+        if interaction.permissions.manage_channels:
             async with self.db_pool.acquire() as conn:
-                await conn.executemany(
-                    "INSERT INTO sticky (channel_id,message_id,message) VALUES ($1,$2,$3);",
-                    [(channel.id, id.id, message)]
+                res = await conn.fetch(
+                    "SELECT channel_id FROM sticky WHERE channel_id = $1", channel.id
                 )
+            if not res:
+                target = self.bot.get_channel(channel.id)
+                id = await target.send(message)
+            
+                async with self.db_pool.acquire() as conn:
+                    await conn.executemany(
+                        "INSERT INTO sticky (channel_id,message_id,message) VALUES ($1,$2,$3);",
+                        [(channel.id, id.id, message)]
+                 )
 
-            embed = discord.Embed(
-                color=discord.Color.green(),
-                title='✅ Sticky message successfully given',
-                description=f"Berhasil menambahkan sticky message pada channel {channel.name}",
-                timestamp=datetime.now()
-            )
+                embed = discord.Embed(
+                    color=discord.Color.green(),
+                    title='✅ Sticky message successfully given',
+                    description=f"Berhasil menambahkan sticky message pada channel {channel.name}",
+                    timestamp=datetime.now()
+                )
+            else:
+                embed = discord.Embed(
+                    color=discord.Color.red(),
+                    title='❌ Sticky message already exist',
+                    description=f"Sticky message telah terpasang pada channel {channel.name}",
+                    timestamp=datetime.now()
+                )
         else:
             embed = discord.Embed(
                 color=discord.Color.red(),
-                title='❎ Sticky message already exist',
-                description=f"Sticky message telah terpasang pada channel {channel.name}",
+                title="❌ You Don't Have Permission To Create Sticky Message",
+                description=f"Permission Manage Channel Dibutuhkan",
                 timestamp=datetime.now()
             )
+
+        embed.set_footer(
+            text=f"{str(interaction.user)}",
+            icon_url=interaction.user.display_avatar.url
+        )
 
         await interaction.followup.send(embed=embed)
     
@@ -75,34 +89,47 @@ class Sticky(commands.GroupCog, group_name="sticky"):
         channel: app_commands.AppCommandChannel
     ) -> None:
         await interaction.response.defer()
-        async with self.db_pool.acquire() as conn:
-            res = await conn.fetch(
-                "SELECT channel_id,message_id FROM sticky WHERE channel_id = $1", channel.id
-            )
-        if not res:
+        if interaction.permissions.manage_channels:
+            async with self.db_pool.acquire() as conn:
+                res = await conn.fetch(
+                    "SELECT channel_id,message_id FROM sticky WHERE channel_id = $1", channel.id
+                )
+            if not res:
+                embed = discord.Embed(
+                    color=discord.Color.red(),
+                    title='❌ Sticky message not exist',
+                    description=f"Tidak ada sticky message pada channel {channel.name}",
+                    timestamp=datetime.now()
+                )
+            else:
+                data = dict(res[0])
+
+                sticky = await channel.fetch()
+                await sticky.delete_messages([discord.Object(id=data['message_id'])])
+
+                async with self.db_pool.acquire() as conn:
+                    await conn.execute(
+                        "DELETE FROM sticky WHERE channel_id = $1;", channel.id
+                    )
+
+                embed = discord.Embed(
+                    color=discord.Color.green(),
+                    title='✅ Sticky message removed successfully',
+                    description=f"Berhasil menghapus sticky message pada channel {channel.name}",
+                    timestamp=datetime.now()
+                )
+        else:
             embed = discord.Embed(
                 color=discord.Color.red(),
-                title='❎ Sticky message not exist',
-                description=f"Tidak ada sticky message pada channel {channel.name}",
+                title="❌ You Don't Have Permission To Delete Sticky Message",
+                description=f"Permission Manage Channel Dibutuhkan",
                 timestamp=datetime.now()
             )
-        else:
-            data = dict(res[0])
 
-            sticky = await channel.fetch()
-            await sticky.delete_messages([discord.Object(id=data['message_id'])])
-
-            async with self.db_pool.acquire() as conn:
-                await conn.execute(
-                    "DELETE FROM sticky WHERE channel_id = $1;", channel.id
-                )
-
-            embed = discord.Embed(
-                color=discord.Color.green(),
-                title='✅ Sticky message removed successfully',
-                description=f"Berhasil menghapus sticky message pada channel {channel.name}",
-                timestamp=datetime.now()
-            )
+        embed.set_footer(
+            text=f"{str(interaction.user)}",
+            icon_url=interaction.user.display_avatar.url
+        )
 
         await interaction.followup.send(embed=embed)
